Create XRPL message log classes
I plan the commit log to simulate the original tree structure internally. It will appear as if it is a linear array holding the longest chain of committed ledgers but internally the tree has to be kept also. Or could this be done by communicating with the actual tree in the replica class?